refactor(auth): realocação e renomeação de arquivos e funções / authUsuario removido
