+++ src/features/Repertorios/Schemas/CitacaoSchema.ts
@@ -9,7 +9,10 @@ import {
   genericSuccessResponse,
   idValidation,
 } from '../../../shared/Validations';
-import { authMiddleware, optionalAuthMiddleware } from '../../Auth/Plugins';
+import {
+  authMiddleware,
+  optionalAuthMiddleware,
+} from '../../../shared/middlewares/Authentication';
 
 export const CitacaoSchema: EntitySchema = {
   get: {
